# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ChecklistItemTemplateRead": {
        "description": "Un item du formulaire de contrôle. `ordre` = ordre d'affichage ; `categorie` = son\nregroupement (`exterieur|interieur|mecanique|documents|securite`, plan.md § 5.1) — le front\ngroupe côté client en triant par `ordre` puis en regroupant par `categorie`, sans dupliquer\nce référentiel.",
        "properties": {
          "categorie": {
            "title": "Categorie",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_required": {
            "title": "Is Required",
            "type": "boolean"
          },
          "libelle": {
            "title": "Libelle",
            "type": "string"
          },
          "ordre": {
            "title": "Ordre",
            "type": "integer"
          },
          "response_type": {
            "title": "Response Type",
            "type": "string"
          },
          "template_id": {
            "format": "uuid",
            "title": "Template Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "template_id",
          "code",
          "libelle",
          "categorie",
          "ordre",
          "is_required",
          "response_type"
        ],
        "title": "ChecklistItemTemplateRead",
        "type": "object"
      },
      "ChecklistTemplateBrief": {
        "description": "`GET /checklist-templates` — sans les items, pour lister les modèles disponibles.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "libelle": {
            "title": "Libelle",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "code",
          "libelle",
          "version",
          "is_active"
        ],
        "title": "ChecklistTemplateBrief",
        "type": "object"
      },
      "ChecklistTemplateRead": {
        "description": "`GET /checklist-templates/{id}` — items triés par `ordre` (garanti par la relation ORM,\n`app/models/checklist.py`).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChecklistItemTemplateRead"
            },
            "title": "Items",
            "type": "array"
          },
          "libelle": {
            "title": "Libelle",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "code",
          "libelle",
          "version",
          "is_active"
        ],
        "title": "ChecklistTemplateRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InspectionCreate": {
        "description": "Idempotent par `client_uuid`, généré côté client (décision C : le brouillon doit pouvoir\nnaître hors ligne). `vehicle_id` doit être en `CONTROLE_EN_COURS` avec une mission active\naffectée à l'appelant.",
        "properties": {
          "client_uuid": {
            "format": "uuid",
            "title": "Client Uuid",
            "type": "string"
          },
          "template_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "vehicle_id": {
            "format": "uuid",
            "title": "Vehicle Id",
            "type": "string"
          }
        },
        "required": [
          "client_uuid",
          "vehicle_id"
        ],
        "title": "InspectionCreate",
        "type": "object"
      },
      "InspectionItemRead": {
        "properties": {
          "commentaire": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commentaire"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "item_template_id": {
            "format": "uuid",
            "title": "Item Template Id",
            "type": "string"
          },
          "photo_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Id"
          },
          "valeur_bool": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valeur Bool"
          },
          "valeur_note": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valeur Note"
          },
          "valeur_num": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valeur Num"
          },
          "valeur_texte": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valeur Texte"
          }
        },
        "required": [
          "id",
          "item_template_id",
          "valeur_bool",
          "valeur_note",
          "valeur_texte",
          "valeur_num",
          "commentaire",
          "photo_id"
        ],
        "title": "InspectionItemRead",
        "type": "object"
      },
      "InspectionItemUpsert": {
        "properties": {
          "commentaire": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commentaire"
          },
          "item_template_id": {
            "format": "uuid",
            "title": "Item Template Id",
            "type": "string"
          },
          "photo_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Id"
          },
          "valeur_bool": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valeur Bool"
          },
          "valeur_note": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valeur Note"
          },
          "valeur_num": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valeur Num"
          },
          "valeur_texte": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valeur Texte"
          }
        },
        "required": [
          "item_template_id"
        ],
        "title": "InspectionItemUpsert",
        "type": "object"
      },
      "InspectionItemsUpsertRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InspectionItemUpsert"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "InspectionItemsUpsertRequest",
        "type": "object"
      },
      "InspectionPatch": {
        "description": "Champs de brouillon modifiables tant que `submitted_at` est nul.",
        "properties": {
          "commentaire": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commentaire"
          },
          "etat_general": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Etat General"
          },
          "kilometrage_releve": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kilometrage Releve"
          }
        },
        "title": "InspectionPatch",
        "type": "object"
      },
      "InspectionRead": {
        "properties": {
          "client_uuid": {
            "format": "uuid",
            "title": "Client Uuid",
            "type": "string"
          },
          "commentaire": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commentaire"
          },
          "conclusion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conclusion"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "driver_id": {
            "format": "uuid",
            "title": "Driver Id",
            "type": "string"
          },
          "etat_general": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Etat General"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InspectionItemRead"
            },
            "title": "Items",
            "type": "array"
          },
          "kilometrage_releve": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kilometrage Releve"
          },
          "mission_id": {
            "format": "uuid",
            "title": "Mission Id",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "submitted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "template_id": {
            "format": "uuid",
            "title": "Template Id",
            "type": "string"
          },
          "vehicle_id": {
            "format": "uuid",
            "title": "Vehicle Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "vehicle_id",
          "mission_id",
          "driver_id",
          "template_id",
          "client_uuid",
          "started_at",
          "submitted_at",
          "kilometrage_releve",
          "etat_general",
          "conclusion",
          "commentaire",
          "created_at"
        ],
        "title": "InspectionRead",
        "type": "object"
      },
      "InspectionSubmitRequest": {
        "description": "Corps optionnel de `POST /inspections/{id}/submit` — permet de poser les champs de\nsynthèse en même temps que la soumission plutôt que d'exiger un `PATCH` préalable.",
        "properties": {
          "commentaire": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commentaire"
          },
          "conclusion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conclusion"
          },
          "etat_general": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Etat General"
          },
          "kilometrage_releve": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kilometrage Releve"
          }
        },
        "title": "InspectionSubmitRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Outil interne de gestion d'achat de véhicules d'occasion — démonstration.",
    "title": "Cardan API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/checklist-templates": {
      "get": {
        "description": "Sans les items (volumétrie dérisoire — quelques modèles) : le détail se récupère par\n`GET /checklist-templates/{id}`, typiquement à partir du `template_id` renvoyé par\n`POST /inspections` / `GET /inspections/{id}`.",
        "operationId": "list_checklist_templates_api_v1_checklist_templates_get",
        "parameters": [
          {
            "in": "query",
            "name": "is_active",
            "required": false,
            "schema": {
              "default": true,
              "title": "Is Active",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChecklistTemplateBrief"
                  },
                  "title": "Response List Checklist Templates Api V1 Checklist Templates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Checklist Templates",
        "tags": [
          "checklist-templates"
        ]
      }
    },
    "/api/v1/checklist-templates/{template_id}": {
      "get": {
        "operationId": "get_checklist_template_api_v1_checklist_templates__template_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChecklistTemplateRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Checklist Template",
        "tags": [
          "checklist-templates"
        ]
      }
    },
...
    "/api/v1/inspections": {
      "post": {
        "operationId": "create_inspection_api_v1_inspections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Inspection",
        "tags": [
          "inspections"
        ]
      }
    },
    "/api/v1/inspections/{inspection_id}": {
      "get": {
        "operationId": "get_inspection_api_v1_inspections__inspection_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inspection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Inspection",
        "tags": [
          "inspections"
        ]
      },
      "patch": {
        "operationId": "patch_inspection_api_v1_inspections__inspection_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inspection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectionPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Inspection",
        "tags": [
          "inspections"
        ]
      }
    },
    "/api/v1/inspections/{inspection_id}/items": {
      "put": {
        "operationId": "put_inspection_items_api_v1_inspections__inspection_id__items_put",
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inspection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectionItemsUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Inspection Items",
        "tags": [
          "inspections"
        ]
      }
    },
    "/api/v1/inspections/{inspection_id}/submit": {
      "post": {
        "operationId": "submit_api_v1_inspections__inspection_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inspection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectionSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit",
        "tags": [
          "inspections"
        ]
      }
    },
...
  }
}